Addressed code duplication comment by SonarQube
